% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsq.R
\name{Rsq}
\alias{Rsq}
\alias{Rsq.lm}
\alias{Rsq.mlm}
\alias{Rsq.glm}
\title{Multiple correlation coefficient methods}
\usage{
Rsq(model, ...)

\method{Rsq}{lm}(model, adjust = TRUE, ...)

\method{Rsq}{mlm}(model, adjust = TRUE, ...)

\method{Rsq}{glm}(model, adjust = TRUE, ...)
}
\arguments{
\item{model}{The fitted model object}

\item{...}{currently ignored, may be used in future methods}

\item{adjust}{logical: should the value be bias adjusted?}
}
\value{
Multiple correlation estimates as required
}
\description{
Calculates R^2, by dafault adjusted, for basic models
}
\examples{
.fm <- lm(medv ~ ., Boston)
c(Rsq(.fm), summary(.fm)$adj.r.squared)
c(Rsq(.fm, adjust = FALSE), summary(.fm)$r.squared)
.tst <- glm.nb(Days ~ Sex/(Age + Eth*Lrn), quine)
Rsq(.tst)
Rsq(.tst, FALSE)
rm(.fm, .tst)
}
