% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_BIC.R
\name{step_AIC}
\alias{step_AIC}
\alias{step_BIC}
\alias{step_GIC}
\alias{dropterm}
\title{Stepwise model construction and inspection}
\usage{
step_AIC(object, ..., trace = 0, k = 2)

step_BIC(object, ..., trace = 0, k = max(2, log(nobs(object))))

step_GIC(object, ..., trace = 0, k = (2 + log(nobs(object)))/2)

dropterm(object, ..., sorted = TRUE, test = default_test(object), k)
}
\arguments{
\item{object}{as for \code{\link[MASS]{stepAIC}}}

\item{...}{additional arguments passed on to main function in \code{MASS}}

\item{trace, k}{as for \code{\link[MASS]{stepAIC}}}

\item{sorted, test}{as for \code{\link[MASS]{dropterm}}}
}
\value{
A fitted model object after stepwise refinement, or a data frame.
}
\description{
Front-ends to \code{\link[MASS]{stepAIC}} and \code{\link[MASS]{dropterm}} with changed defaults.
\code{step_BIC} implements a stepwise selection with BIC as the criterion and
\code{step_GIC} uses an experimental criterion with a penalty midway between AIC and BIC: the
"Goldilocks" criterion.
}
\examples{
fm <- glm.nb(Days ~ .^3, quine)
dropterm(fm_aic <- step_AIC(fm))
dropterm(fm_bic <- step_BIC(fm))
}
