\name{WWRUtilities-package}
\alias{WWRUtilities-package}
\alias{WWRUtilities}
\docType{package}
\title{
Support Software and Datasets for WWR
}
\description{
As the title says.  Undocumented here, but explained as
  needed in the workshop.
}
\details{
\tabular{ll}{
Package: \tab WWRUtilities\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-12-30\cr
License: \tab GPL (>= 2)\cr
Imports: \tab lazyData\cr
Suggests: rpart, randomForest, xtable
Depends: \tab graphics, grDevices, utils, stats, MASS, SOAR, splines, lazyData, splines, tidyverse, WWRGraphics\cr
}
See the workshop for details.
}
\author{
Bill Venables

Maintainer: Bill Venables <Bill.Venables@gmail.com>
}

\keyword{ package }
\seealso{
  The packages \code{tidyverse}, \code{MASS}, \code{SOAR} and \code{lazyData}
}

